Fixed long standing bug in vacc
vacc_init_xblock/SyncConst was being set to actual_veclen-2 as a relic from the original gavrt/vacc. That version had an additional delay in the loop around the Counter, so the constant was actual_veclen-2 to account for the latency from reset to zero of the counter (1) and the latency of a comparator (1).

In this version, the latency is only from the reset to zero of the counter, so the constant should be actual_veclen-1, as corrected here.

This caused the valid signal to drift with respect to the accumulated output vector except in the case that a sync signal was asserted more often than the vacc/Counter period.